wip refactor to use classes
