+++ classes/OSMPhotoNote.class.php
@@ -1,0 +1,10 @@
+<?php
+
+/** Represents an OSM note to which a photo is attached */
+class OSMPhotoNote
+{
+    public $note_id;
+    public $status;
+    public $closed_at;
+    public $photo_ids;
+}
