+++ db_helper.php
@@ -16,93 +16,4 @@ class DBHelper
         $this->_connection->close();
     }
 
-    public function createTable()
-    {
-        $this->_connection->query(
-            'CREATE TABLE IF NOT EXISTS photos(
-                file_id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-                file_ext VARCHAR(10) NOT NULL,
-                creation_time DATETIME NOT NULL,
-                note_id BIGINT UNSIGNED
-            )'
-        );
-    }
-
-    public function newPhoto($file_ext)
-    {
-        $stmt = $this->_connection->prepare(
-            'INSERT INTO photos(file_ext, creation_time)
-                VALUES (?, NOW())'
-        );
-        $stmt->bind_param('s', $file_ext);
-        $stmt->execute();
-        return $this->_connection->insert_id;
-    }
-
-    public function deletePhoto($file_id)
-    {
-        $stmt = $this->_connection->prepare('DELETE FROM photos WHERE file_id=?');
-        $stmt->bind_param('i', $file_id);
-        $stmt->execute();
-    }
-
-    public function getOldInactivePhotos()
-    {
-        $stmt = $this->_connection->prepare(
-            'SELECT file_id, file_ext, note_id FROM photos
-                WHERE note_id IS NULL
-                AND creation_time < ADDDATE(NOW(), INTERVAL -? HOUR)'
-        );
-        $max_age = Config::MAX_TMP_LIFETIME_HOURS;
-        $stmt->bind_param('i', $max_age);
-        $stmt->execute();
-        $result = $stmt->get_result()->fetch_all(MYSQLI_ASSOC);
-        return $result;
-    }
-
-    public function getOldestActivePhotos($num)
-    {
-        $stmt = $this->_connection->prepare(
-            'SELECT file_id, file_ext, note_id FROM photos
-                WHERE note_id IS NOT NULL
-                ORDER BY creation_time
-                LIMIT ?'
-        );
-        $stmt->bind_param('i', $num);
-        $stmt->execute();
-        $result = $stmt->get_result()->fetch_all(MYSQLI_ASSOC);
-        return $result;
-    }
-
-    public function getActivePhotos()
-    {
-        $result = $this->_connection->query(
-            'SELECT file_id, file_ext, note_id
-                FROM photos WHERE note_id IS NOT NULL'
-        );
-        return $result->fetch_all(MYSQLI_ASSOC);
-    }
-
-    public function getInactivePhotos($photo_ids)
-    {
-        $in = str_repeat('?,', count($photo_ids) - 1) . '?';
-        $stmt = $this->_connection->prepare(
-            "SELECT file_id, file_ext FROM photos
-                WHERE file_id IN ($in)
-                AND note_id IS NULL"
-        );
-        $stmt->bind_param(str_repeat('i', count($photo_ids)), ...$photo_ids);
-        $stmt->execute();
-        return $stmt->get_result()->fetch_all(MYSQLI_ASSOC);
-    }
-
-    public function activatePhoto($photo_id, $note_id)
-    {
-        $stmt = $this->_connection->prepare(
-            "UPDATE photos SET note_id = ?
-                WHERE file_id = ?"
-        );
-        $stmt->bind_param('ii', $note_id, $photo_id);
-        $stmt->execute();
-    }
 }
